# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MaintSys API",
        "description": "API de gerenciamento de manutenção industrial - MaintSys",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/machine-readings": {
            "get": {
                "tags": [
                    "Leituras de Máquinas"
                ],
                "summary": "Lista todas as leituras de sensores",
                "operationId": "4aa2cd412e243b07875bc66b86f598b0",
                "responses": {
                    "200": {
                        "description": "Lista de leituras de sensores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "machine_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "sensor_key": {
                                                "type": "string",
                                                "example": "temperatura"
                                            },
                                            "value": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 75.5
                                            },
                                            "unit": {
                                                "type": "string",
                                                "example": "°C"
                                            },
                                            "read_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-10T10:30:00Z"
                                            },
                                            "machine": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Leituras de Máquinas"
                ],
                "summary": "Registra uma nova leitura de sensor",
                "operationId": "af9cdca04ccd07023826755bcb1f707d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "machine_id",
                                    "sensor_key",
                                    "value",
                                    "read_at"
                                ],
                                "properties": {
                                    "machine_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sensor_key": {
                                        "type": "string",
                                        "example": "pressao"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3.2
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "bar"
                                    },
                                    "read_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-10T10:30:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Leitura registrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "machine_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "sensor_key": {
                                            "type": "string",
                                            "example": "pressao"
                                        },
                                        "value": {
                                            "type": "number",
                                            "example": 3.2
                                        },
                                        "unit": {
                                            "type": "string",
                                            "example": "bar"
                                        },
                                        "read_at": {
                                            "type": "string",
                                            "example": "2024-01-10T10:30:00Z"
                                        },
                                        "machine": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/machine-readings/{id}": {
            "get": {
                "tags": [
                    "Leituras de Máquinas"
                ],
                "summary": "Retorna uma leitura de sensor específica",
                "operationId": "ae528d42b697f330281ec9147925d626",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da leitura",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da leitura",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "machine_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "sensor_key": {
                                            "type": "string",
                                            "example": "temperatura"
                                        },
                                        "value": {
                                            "type": "number",
                                            "example": 75.5
                                        },
                                        "unit": {
                                            "type": "string",
                                            "example": "°C"
                                        },
                                        "read_at": {
                                            "type": "string",
                                            "example": "2024-01-10T10:30:00Z"
                                        },
                                        "machine": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Leitura não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Leituras de Máquinas"
                ],
                "summary": "Atualiza uma leitura de sensor",
                "operationId": "65f8cda109da14dfb15b0a14a47d58e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da leitura",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sensor_key": {
                                        "type": "string",
                                        "example": "temperatura"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 80
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "°C"
                                    },
                                    "read_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-10T11:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Leitura atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "sensor_key": {
                                            "type": "string",
                                            "example": "temperatura"
                                        },
                                        "value": {
                                            "type": "number",
                                            "example": 80
                                        },
                                        "unit": {
                                            "type": "string",
                                            "example": "°C"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Leitura não encontrada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Leituras de Máquinas"
                ],
                "summary": "Remove uma leitura de sensor",
                "operationId": "77f7ab88daa11002a4455ac4ab3a2b11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da leitura",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leitura removida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Leitura removida"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Leitura não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}